Add Mappings API environment variables (PHNX-2414) (#65)
Motivation
------------
Importing waivers from Webstore 2 will require getting the SiteId for a business entity from the Mappings service and we need to add the mappings load balancer to an environment variable in the deployment pipelines used by Spinnaker.

Modifications
---------------
Added environment variables to the production and smoke test pipeline templates for the mapping service api.

https://centeredge.atlassian.net/browse/PHNX-2414