Add Shawarma support to microservices (ARCH-56)
Motivation
----------
Use a single variable on the pipeline template to apply the Shawarma
annotations to deployed pods.

Modifications
-------------
Added "shawarmaEnabled" and "shawarmaStaging" variables to the
microservice template, which automatically adds the annotations.

Set the Waivers microservice to have Shawarma enabled, and also updated
the special configuration for the Jobs processor to apply the Shawarma
annotation.