Update liveness probes to use ready checks (PHNX-3284) (#105)
Motivation
------------
Kubernetes will only restart a pod which is failing the liveness probe, and not a pod which is failing the readiness probe.  We currently have the liveness probes configured to hit `/health` where the readiness probes hit `/health?type=ready`.  The `/health` endpoint does not attempt to retrieve a document from couchbase and will pass even when couchbase cannot be accessed which leaves us with pods that cannot process any real requests.

Modifications
---------------
Changed all templates liveness probes to use `/health?type=ready` to check for liveness.

https://centeredge.atlassian.net/browse/PHNX-3284