Increase memory limits for business entities (PHNX-6016)
Motivation
------------
We had an outage of the business entities microservice due to out of memory conditions causing the pods to be killed by k8s.  We increased the memory limits in production to get things back up and running and we need to put those changes in source control now.  We also want to remove the nginx proxy container from production as it is no longer needed.

Modifications
---------------
- Increased memory limits in business entities config
- Removed nginx proxy container from producution pipelines