Update proxy pipeline template to use AdvantageAuthProxy (PHNX-6738)
Motivation
----------
We'd like to use an authenticating proxy rather than a plain Nginx proxy
for talking to Advantage. This newer proxy has a lot of value added
functionality.

Modifications
-------------
Convert the existing pipeline template from JSON to YAML for legibility.

Replace the Nginx artifact and container with the auth proxy, switching
out environment variables as required.

Add a DataDog metric and change the HPA to scale based on that metric
rather than CPU. Add env vars to the auth container to ensure that the
metrics are always posted correctly to match with the metric query.