Create prod pipeline template with metric scaling (PHNX-6702) (#168)
Motivation
----------
CPU scaling isn't really cutting it for us.

Modifications
-------------
Convert the existing pipeline template to YAML for legibiilty as a
starting point.

Control metric tagging using environment variables across staging, prod,
and prod jobs to ensure it matches with the metric query. Use the
Spinnaker app name as the "app" tag, "production" or "staging" as the
environment, and "api" or "jobs" as the "detail" tag.

Include the DataDog query metric and update the HPA to scale based on
this metric for the prod API and Jobs deploys. Scale the API based on
incoming HTTP requests, and scale the Jobs based on incoming message
queue items being processed.

Introduce new variables which optionally control the scaling thresholds
used for both API and Jobs scaling.

Remove "?type=health" from the healh checks as this doesn't do anything
with the new health checks.

Drop the Docker trigger from the template. Instead, each created
pipeline should be configured to trigger from the Jenkins job on the
main branch.

Results
-------
Each microservice may be converted to this template one by one.

https://centeredge.atlassian.net/browse/PHNX-6702